internal/storage: rework label index
Since we use single rbf store, there is no need to design for concurrent
access. This simplifies reasoning for labels index.

We only need to know where the label was found. So we have a one to many relation
between labels and metrics id.

id => [label1,label2 ...]. This works better with mutex encoding we know labels
are scoped per column which lowers cardinality. We can support up to 4,294,967,295
unique values per label name.

For prometheus systems cardinality is already a sensitive topic , we expect this
design to cover fully existing tsdb use cases.

The biggest challenge of mutex columns is database size, we expect the size to grow
much faster than when we were using BSI.

The biggest advantage is we can just use bitmaps.Row call to select series based one
matchers.